global: restructure workflow run managers code
